+++ src/Starlight/Actor.hs
@@ -30,34 +30,34 @@ import Unit.Mass
 import Unit.Time
 
 data Actor = Actor
-  { position  :: !(V3 (Mega Metres Float))
-  , velocity  :: !(V3 ((Mega Metres :/: Seconds) Float))
-  , rotation  :: !(Quaternion Float)
-  , mass      :: !(Kilo Grams Float)
-  , magnitude :: !(Mega Metres Float) -- approx. equivalent to diameter; should bound the actor’s geometry
+  { position  :: !(V3 (Mega Metres Double))
+  , velocity  :: !(V3 ((Mega Metres :/: Seconds) Double))
+  , rotation  :: !(Quaternion Double)
+  , mass      :: !(Kilo Grams Double)
+  , magnitude :: !(Mega Metres Double) -- approx. equivalent to diameter; should bound the actor’s geometry
   }
   deriving (Generic, Show)
 
-position_ :: HasCallStack => Lens' Actor (V3 (Mega Metres Float))
+position_ :: HasCallStack => Lens' Actor (V3 (Mega Metres Double))
 position_ = field @"position".asserting (none isNaN)
 
-velocity_ :: HasCallStack => Lens' Actor (V3 ((Mega Metres :/: Seconds) Float))
+velocity_ :: HasCallStack => Lens' Actor (V3 ((Mega Metres :/: Seconds) Double))
 velocity_ = field @"velocity".asserting (none isNaN)
 
-rotation_ :: HasCallStack => Lens' Actor (Quaternion Float)
+rotation_ :: HasCallStack => Lens' Actor (Quaternion Double)
 rotation_ = field @"rotation".asserting (none isNaN)
 
-mass_ :: HasCallStack => Lens' Actor (Kilo Grams Float)
+mass_ :: HasCallStack => Lens' Actor (Kilo Grams Double)
 mass_ = field @"mass".asserting (not.isNaN)
 
-magnitude_ :: HasCallStack => Lens' Actor (Mega Metres Float)
+magnitude_ :: HasCallStack => Lens' Actor (Mega Metres Double)
 magnitude_ = field @"magnitude".asserting (not.isNaN)
 
 
-transformToActor :: Actor -> Transform Float (Mega Metres) (Mega Metres)
+transformToActor :: Actor -> Transform Double (Mega Metres) (Mega Metres)
 transformToActor Actor{ position, rotation } = mkTranslation (prj <$> position) >>> mkRotation rotation
 
-applyForce :: HasCallStack => V3 ((Kilo Grams :*: Mega Metres :/: Seconds :/: Seconds) Float) -> Seconds Float -> Actor -> Actor
+applyForce :: HasCallStack => V3 ((Kilo Grams :*: Mega Metres :/: Seconds :/: Seconds) Double) -> Seconds Double -> Actor -> Actor
 applyForce force dt a = a & velocity_ +~ ((.*. dt) . (./. a^.mass_) <$> force)
