+++ src/Starlight/View.hs
@@ -34,9 +34,9 @@ import Unit.Length
 data View = View
   { ratio :: Int    -- ^ Ratio of window pixels per context pixel.
   , size  :: V2 (Window.Pixels Int)
-  , zoom  :: Float
-  , scale :: Float
-  , focus :: V2 (Mega Metres Float)
+  , zoom  :: Double
+  , scale :: Double
+  , focus :: V2 (Mega Metres Double)
   }
 
 aspectRatio :: View -> Float
@@ -46,7 +46,7 @@ aspectRatio View{ size } = size'^._x / size'^._y where
 deviceSize :: View -> V2 (Context.Pixels Int)
 deviceSize View{ ratio, size } = fmap Context.Pixels (ratio *^ fmap Window.getPixels size)
 
-lengthToWindowPixels :: View -> Float
+lengthToWindowPixels :: View -> Double
 lengthToWindowPixels View{ zoom, scale } = 1/zoom * scale
 
 
@@ -54,25 +54,25 @@ newtype ClipUnits a = ClipUnits { getClipUnits :: a }
 newtype Zoomed a = Zoomed { getZoomed :: a }
 
 
-toContext :: View -> Transform Float ClipUnits Context.Pixels
+toContext :: View -> Transform Double ClipUnits Context.Pixels
 toContext View{ size } = mkScale (pure 1 & _xy .~ 1 / (fromIntegral <$> size))
 
-toWindow :: View -> Transform Float Context.Pixels Window.Pixels
+toWindow :: View -> Transform Double Context.Pixels Window.Pixels
 toWindow View{ ratio } = mkScale (pure 1 & _xy .~ fromIntegral ratio)
 
-toZoomed :: View -> Transform Float Window.Pixels Zoomed
+toZoomed :: View -> Transform Double Window.Pixels Zoomed
 toZoomed View{ zoom } = mkScale (pure 1 & _xy .~ pure (1/zoom))
 
-toSystem :: View -> Transform Float Zoomed (Mega Metres)
+toSystem :: View -> Transform Double Zoomed (Mega Metres)
 toSystem View{ scale, focus } = mkScale (pure scale) >>> mkTranslation (ext (prj <$> negated focus) 0)
 
-transformToWindow :: View -> Transform Float ClipUnits Window.Pixels
+transformToWindow :: View -> Transform Double ClipUnits Window.Pixels
 transformToWindow view = toContext view >>> toWindow view
 
-transformToZoomed :: View -> Transform Float ClipUnits Zoomed
+transformToZoomed :: View -> Transform Double ClipUnits Zoomed
 transformToZoomed view = transformToWindow view >>> toZoomed view
 
-transformToSystem :: View -> Transform Float ClipUnits (Mega Metres)
+transformToSystem :: View -> Transform Double ClipUnits (Mega Metres)
 transformToSystem view = transformToZoomed view >>> toSystem view
